# rest/src/main/resources/io/jmix/rest/rest-openapi.yaml
openapi: 3.0.1
info:
  title: Jmix Generic REST API
  description: |2

    ## Overview

    Jmix Generic REST API provides the following functionality:

    * Entities CRUD operations
    * Execution of predefined JPQL queries
    * Execution of exposed service methods
    * Getting metadata (entities, fetch plans, enumerations, datatypes)
    * Getting localized messages
    * Getting current user permissions (roles, access to entities and attributes, specific policies)
    * Getting attributes of the current user
    * Working with files

    The REST API is available at the `/rest/` URL, e.g. `http://localhost:8080/rest/entities/Order`

    ## Authentication

    ### Authentication using the OAuth2 protocol

    Jmix REST API uses an OAuth2 resource owner password credentials grant to protect the endpoints (https://datatracker.ietf.org/doc/html/rfc6749#section-4.3).

    Before accessing a protected resource, a user must get an authentication token. The token can be obtained at the `/oauth/token` URL.

    The `/oauth/token` endpoint is protected with the basic authentication. The request must contain the `Authorization` header for basic
    authentication with REST API client credentials. Please note that client credentials are not the same as the user credentials.
    Client means an application that uses the REST API. The default client credentials are defined by the following
...
    The request format must be `application/x-www-form-urlencoded`.

...

...

    ```
...
    ```

...

...

...
servers:
...
security:
  - oauth2:
      - rest-api
...
paths:
...
  /entities/{entityName}:
    get:
      tags:
        - Entities
      summary: Get a list of entities
      description: |
        Gets a list of entities
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: fetchPlan
          in: query
          description: Name of the fetch plan which is used for loading the entity
          schema:
            type: string
        - name: limit
          in: query
          description: Number of extracted entities
          schema:
            type: string
        - name: offset
          in: query
          description: Position of the first result to retrieve
          schema:
            type: string
        - name: sort
          in: query
          description: |
            Name of the field to be sorted by. If the name is preceeding by the '+' character, then
            the sort order is ascending, if by the '-' character then descending. If there is no special
            character before the property name, then ascending sort will be used.
          schema:
            type: string
        - name: returnNulls
          in: query
          description: Specifies whether null fields will be written to the result JSON
          schema:
            type: boolean
        - name: returnCount
          in: query
          description: Specifies whether the total count of entities should be returned
            in the 'X-Total-Count' header
          schema:
            type: boolean
        - name: dynamicAttributes
          in: query
          description: Specifies whether entity dynamic attributes should be returned
          schema:
            type: boolean
      responses:
        200:
          description: Success. The list of entities is returned in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entityFull'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Not found. MetaClass for the entity with the given name not
            found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      tags:
        - Entities
      summary: Create new entity
      description: |
        Creates new entity. The method expects a JSON with entity object in the request body. The entity object
        may contain references to other entities. These references are processed according to the following rules:

        * If the entity property is annotated with @Composition in the entity java class, then it will be saved with the
        main entity
        * Otherwise a referenced entity with the given id will be searched. If it is found then the saved entity will
        have a reference to it. Otherwise, a response with code 400 will be returned.
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
      requestBody:
        description: |
          A JSON object with the entity
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity'
        required: false
      responses:
        201:
          description: Entity created. The created entity is returned in the response
            body.
          headers:
            location:
              description: A URL the created object can be accessed with
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entityFull'
        400:
          description: Bad request. For example, the entity may have a reference to
            the non-existing entity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        403:
          description: Forbidden. The user doesn't have permissions to create the
            entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Not found. MetaClass for the entity with the given name not
            found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      x-codegen-request-body-name: entityJson
  /entities/{entityName}/{entityId}:
    get:
      tags:
        - Entities
      summary: Get a single entity by id
      description: |
        Gets a single entity by identifier
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: entityId
          in: path
          description: Entity identifier
          required: true
          schema:
            type: string
        - name: fetchPlan
          in: query
          description: Name of the fetch plan which is used for loading the entity
          schema:
            type: string
        - name: returnNulls
          in: query
          description: Specifies whether null fields will be written to the result JSON
          schema:
            type: boolean
        - name: dynamicAttributes
          in: query
          description: Specifies whether entity dynamic attributes should be returned
          schema:
            type: boolean
      responses:
        200:
          description: Success. The entity is returned in the response body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entityFull'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: |
            MetaClass not found or entity with the five identifier not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      tags:
        - Entities
      summary: Update the entity
      description: |
        Updates the entity. Only fields that are passed in the JSON object (the request body) are updated.
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: entityId
          in: path
          description: Entity identifier
          required: true
          schema:
            type: string
      requestBody:
        description: JSON object with the entity
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity'
        required: false
      responses:
        200:
          description: Success. The updated entity is returned in the response body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entityFull'
        403:
          description: Forbidden. The user doesn't have permissions to update the
            entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: |
            MetaClass not found or entity with the given identifier not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      x-codegen-request-body-name: entityJson
    delete:
      tags:
        - Entities
      summary: Delete the entity
      description: Deletes the entity
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: entityId
          in: path
          description: Entity identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success. Entity was deleted.
          content: {}
        403:
          description: Forbidden. The user doesn't have permissions to delete the
            entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: |
            MetaClass not found or entity with the given identifier not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /entities/{entityName}/search:
    get:
      tags:
        - Entities
      summary: Find entities by filter conditions
      description: |2

        Finds entities by filter conditions. The filter is defined by JSON object that is passed as in URL parameter.

        Filter JSON example:

        ```json
        {
            "conditions": [
                {
                    "group": "OR",
                    "conditions": [
                        {
                            "property": "stringField",
                            "operator": "startsWith",
                            "value": "ABC"
                        },
                        {
                            "property": "relatedEntity.intField",
                            "operator": ">",
                            "value": 100
                        }
                    ]
                },
                {
                    "property": "booleanField",
                    "operator": "=",
                    "value": true
                }
            ]
        }
        ```

        Filter supports OR and AND grouping. All conditions in the root "conditions" field are automatically added to the AND group.

        Conditions for properties that are references to other entities must have values with string representation of
        entitiy primary key.

        If the operator is "in" or "notIn" then the value should be a JSON array.

        ```
        {
            "conditions": [
                {
                    "property": "linkedEntityField",
                    "operator": "=",
                    "value": "2de6a78f-7bef-89a7-eb5e-b725582f23af"
                },
                {
                  "property": "intField",
                  "operator": "in",
                  "value": [100, 200, 300]
                }
            ]
        }
        ```

        Supported operators:

        * =
        * &lt;&gt;
        * &lt;
        * &lt;=
        * &gt;
        * &gt;=
        * startsWith
        * endsWith
        * contains
        * notEmpty
        * in
        * notIn
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
        - name: filter
          in: query
          description: JSON with filter definition
          schema:
            type: string
        - name: fetchPlan
          in: query
          description: Name of the fetch plan which is used for loading the entity
          schema:
            type: string
        - name: limit
          in: query
          description: Number of extracted entities
          schema:
            type: string
        - name: offset
          in: query
          description: Position of the first result to retrieve
          schema:
            type: string
        - name: sort
          in: query
          description: |
            Name of the field to be sorted by. If the name is preceeding by the '+' character, then
            the sort order is ascending, if by the '-' character then descending. If there is no special
            character before the property name, then ascending sort will be used.
          schema:
            type: string
        - name: returnNulls
          in: query
          description: Specifies whether null fields will be written to the result JSON
          schema:
            type: boolean
        - name: returnCount
          in: query
          description: Specifies whether the total count of entities should be returned
            in the 'X-Total-Count' header
          schema:
            type: boolean
        - name: dynamicAttributes
          in: query
          description: Specifies whether entity dynamic attributes should be returned
          schema:
            type: boolean
      responses:
        200:
          description: Success. Entities that conforms filter conditions are returned
            in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entityFull'
        400:
          description: Bad request. For example, the condition value cannot be parsed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Not found. MetaClass for the entity with the given name not
            found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      tags:
        - Entities
      summary: Find entities by filter conditions
      description: |2

        Finds entities by filter conditions. Works like the corresponding GET method, but filter definition and all request parameters are
        passed in the request body as JSON object. All possible parameters ("fetchPlan", "sort", "returnNulls", etc.) are the same as in the
        GET version of the `/search` operation.

        Request body example:

        ```json
        {
            "filter": {
              "conditions": [
                  {
                      "group": "OR",
                      "conditions": [
                          {
                              "property": "stringField",
                              "operator": "startsWith",
                              "value": "ABC"
                          },
                          {
                              "property": "relatedEntity.intField",
                              "operator": ">",
                              "value": 100
                          }
                      ]
                  },
                  {
                      "property": "booleanField",
                      "operator": "=",
                      "value": true
                  }
              ]
            },
            "fetchPlan": "someFetchPlanName",
            "returnCount": true
        }
        ```
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success. Entities that conforms filter conditions are returned
            in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entityFull'
        400:
          description: Bad request. For example, the condition value cannot be parsed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        403:
          description: Forbidden. The user doesn't have permissions to read the entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Not found. MetaClass for the entity with the given name not
            found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /metadata/entities/{entityName}:
    get:
      tags:
        - Metadata
      summary: Get entity metadata
      description: Gets entity metadata
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entityMetadata'
        404:
          description: |
            MetaClass not found or query with the given name not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /messages/entities/{entityName}:
    get:
      tags:
        - Messages
      summary: Get localized messages for entity and its attributes
      description: |
        Method gets localized messages for entity and its attributes. It returns a JSON object that represents
        a map of localized messages. The key is entity or entity property name, the value is localized message.
      parameters:
        - name: entityName
          in: path
          description: Entity name
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
        404:
          description: |
            MetaClass not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
...
    error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error description
...
    entity:
      type: object
      properties:
        id:
          type: string
          example: 4d8ef1d4-90b6-c075-8bfd-79181ff889ef
        otherEntityFields:
          type: string
          example: There must be a set of entity fields instead of this one here
    entityFull:
      required:
        - _entityName
        - id
      type: object
      properties:
        id:
          type: string
          example: 4d8ef1d4-90b6-c075-8bfd-79181ff889ef
        _entityName:
          type: string
          description: Entity name
          example: sec$User
        _instanceName:
          type: string
          description: Entity instance name. This value must be used for displaying
            the entity in list.
          example: Smith [smith]
        otherEntityFields:
          type: string
          description: There must be a set of entity fields instead of this one here.
          example: There must be a set of entity fields instead of this one here
...
    entityMetadata:
      type: object
      properties:
        entityName:
          type: string
          example: ref$Driver
        ancestor:
          type: string
          example: sys$StandardEntity
        properties:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: name
              attributeType:
                type: string
                enum:
                  - DATATYPE
                  - ENUM
                  - ASSOCIATION
                  - COMPOSITION
              type:
                type: string
                description: |
                  For a simple datatype there will be a datatype name here (string, int, decimal, boolean, date, dateTime, etc.)
                  You can find all simple datatype names in the source code of the Datatypes class in the systemDatatypeNames property.
                  If the property is the link to the entity then an entity name will be placed to this field, e.g. ref$Colour
                example: string
              cardinality:
                type: string
                enum:
                  - NONE
                  - ONE_TO_ONE
                  - MANY_TO_ONE
                  - ONE_TO_MANY
                  - MANY_TO_MANY
              mandatory:
                type: boolean
              readOnly:
                type: boolean
              persistent:
                type: boolean
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          tokenUrl: http://localhost:8080/rest/oauth/token
          scopes:
            rest-api: use REST API
    tokenEndpointBasic:
      type: http
      scheme: basic